Port rebar_port_compiler to Windows
This is the first attempt on the task. Can compile simple driver
(ebloom), but errors and warning messages are lost because a new
window opens for the compile process (cmd /c ... ?) and closes it
when compiler exits. Must also try an approach where environment
variables from the command line are fully expanded (no need for
'cmd /c' then, hopefully).